Separate ICRC-112 logic to different service
